# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tubi API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/tasks/evaluate": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Rate",
        "operationId": "rate_api_tasks_evaluate_post",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Tasks",
        "operationId": "tasks_api_tasks_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "analytics",
                    "automation",
                    "education",
                    "marketing",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "working",
                    "ready",
                    "priority"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "score_desc",
                "newest"
              ],
              "type": "string",
              "default": "score_desc",
              "title": "Sort"
            }
          },
          {
            "name": "include_drafts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Drafts"
            }
          },
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_Task_"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Task",
        "operationId": "task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/tasks/{task_id}/publish": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Publish",
        "operationId": "publish_api_tasks__task_id__publish_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "accept-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Breakdown": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "earned": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Earned"
          },
          "max": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Max"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "earned",
          "max"
        ],
        "title": "Breakdown"
      },
      "Card": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "analytics",
              "automation",
              "education",
              "marketing",
              "other"
            ],
            "title": "Category",
            "default": "analytics"
          },
          "context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "need": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Need"
          },
          "users": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Users"
          },
          "data": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "expected_result": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Result"
          },
          "success_criteria": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Criteria"
          },
          "constraints": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          },
          "interaction_format": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Format"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Card"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "fields"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "EvaluateInput": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "card"
        ],
        "title": "EvaluateInput"
      },
...
      "Improvement": {
        "properties": {
          "fields": {
            "items": {
              "type": "string",
              "enum": [
                "title",
                "context",
                "need",
                "users",
                "data",
                "expected_result",
                "success_criteria",
                "constraints",
                "contact",
                "interaction_format"
              ]
            },
            "type": "array",
            "title": "Fields"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fields",
          "message"
        ],
        "title": "Improvement"
      },
...
      "Items_Task_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Items[Task]"
      },
...
      "Rating": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "level": {
            "type": "string",
            "enum": [
              "draft",
              "working",
              "ready",
              "priority"
            ],
            "title": "Level"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array",
            "title": "Breakdown"
          },
          "improvements": {
            "items": {
              "$ref": "#/components/schemas/Improvement"
            },
            "type": "array",
            "title": "Improvements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "score",
          "level",
          "breakdown",
          "improvements"
        ],
        "title": "Rating"
      },
      "SaveInput": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "card",
          "confirmed"
        ],
        "title": "SaveInput"
      },
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "card",
          "rating",
          "confirmed",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Task"
      },
...
    }
  }
}